// src/types/account.ts
import { Field } from "./auth";

export type Account = {
  id: string;
  account_id: string;
  username: string;
  acct: string;
  display_name: string;
  locked: boolean;
  domain: string | null;
  bot: boolean;
  discoverable: boolean;
  hide_collections: boolean;
  group: boolean;
  created_at: string;
  note: string;
  uri: string;
  url: string;
  avatar: string;
  image_url: string;
  avatar_static: string;
  header: string;
  header_static: string;
  primary_community_slug: string;
  primary_community_name: string;
  followers_count: number;
  following_count: number;
  statuses_count: number;
  last_status_at: string;
  collection_count: number;
  community_count: number;
  country: string;
  country_common_name: string;
  dob: string;
  is_followed: boolean;
  is_requested: boolean;
  subtitle: string;
  contributor_role: string;
  voices: string;
  media: string;
  hash_tag_count: number;
  noindex: boolean;
  emojis: [];
  fields: Field[];
  tags: [];
  email: string;
  phone: string;
  about_me: string;
  source?: {
    privacy: string;
    sensitive: string;
    language: null;
    note: string;
    fields: [];
    follow_requests_count: number;
    hide_collections?: string;
    discoverable?: boolean;
    indexable: boolean;
    email?: string;
    phone?: string;
  };
};